Remove Folder and fileSystemEntity, rename File to FileUtility
