+++ src/main/java/edu/chl/proton/model/Workspace.java
@@ -1,6 +1,7 @@
 package edu.chl.proton.model;
 
-import java.io.IOException;
+import java.io.*;
+import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -11,7 +12,7 @@ import java.util.List;
 public class Workspace implements IFileHandler, IDocumentHandler {
     private List<Document> tabs = new ArrayList<>();
     private Document currentDocument;
-    private Folder currentDirectory;
+    private java.io.File currentDirectory;
     private DocumentFactory factory = new DocumentFactory();
 
     public Workspace() {
@@ -31,8 +32,12 @@ public class Workspace implements IFileHandler, IDocumentHandler {
        currentDocument.save();
     }
 
-    public void setCurrentDirectory(Folder folder) {
-        currentDirectory = folder;
+    public void setCurrentDirectory(File directory) throws  IOException {
+        if(!directory.isDirectory()) {
+            throw new IOException("Trying to set a file as directory");
+        }
+
+        currentDirectory = directory;
     }
 
     public String getCurrentDirectory() {
@@ -66,16 +71,11 @@ public class Workspace implements IFileHandler, IDocumentHandler {
 
     }
 
-    @Override
-    public String getDirectory() {
-        return null;
-    }
-
-    public void setDirectory(Folder folder) {
+    public void setDirectory(File folder) {
         currentDirectory = folder;
     }
 
-    public Folder getDirectory(Folder folder) {
+    public File getDirectory() {
         return currentDirectory;
     }
